Fixed an issue in which M (the counter variable for the Simpsons Rule) wasn't always even. In the event that M is not even, increment M by 1 and readjust the value for dt used by the algorithm. Further, I've renamed the folder since I killed the old files.

+++ QNB1/src/main.f95
@@ -40,7 +40,7 @@ program main
     read(10,nml=params)
     close(10)
 
-    N = int(time_limit / dt1)
+    N = nint(time_limit / dt1)
     allocate(rho(0:N,S,S)) !Sets up the storage for the data points
     rho(0,:,:) = rho0 ! Initialize the storage
